# API_SPECIFICATION.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "iOS Job Matching Backend API",
    "version": "1.0.0",
    "description": "Complete API specification for iOS job matching application backend"
  },
  "servers": [
    {
      "url": "https://birjobbackend-ir3e.onrender.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/users/analytics/{device_id}": {
      "get": {
        "summary": "Get user analytics and insights",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "User analytics data",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "profileInsights": {
                      "profileStrength": 85,
                      "profileCompleteness": 90,
                      "marketFit": 78
                    },
                    "jobActivity": {
                      "totalJobsViewed": 156,
                      "totalJobsSaved": 23,
                      "totalApplications": 8
                    },
                    "matchingInsights": {
                      "totalMatches": 89,
                      "averageMatchScore": 76
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/jobs/by-company": {
      "get": {
        "summary": "Get top companies by job count from current scraping cycle",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of companies to return",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Top companies statistics",
            "content": {
              "application/json": {
                "example": {
                  "companies": [
                    {
                      "company": "ABB",
                      "job_count": 131,
                      "first_job": "2025-06-15T15:25:34.595587",
                      "latest_job": "2025-06-15T15:25:34.595587"
                    }
                  ],
                  "limit": 20,
                  "data_freshness": "current_cycle_only",
                  "note": "All data from current scraping cycle (refreshed every 4-5 hours)",
                  "timestamp": "2025-06-15T16:15:33.760485"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/jobs/keywords": {
      "get": {
        "summary": "Get most popular keywords in job titles from current scraping cycle",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of keywords to return",
            "schema": {
              "type": "integer",
              "minimum": 10,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Popular keywords statistics",
            "content": {
              "application/json": {
                "example": {
                  "keywords": [
                    {
                      "keyword": "mütəxəssis",
                      "frequency": 548,
                      "percentage": 17.0
                    },
                    {
                      "keyword": "engineer",
                      "frequency": 191,
                      "percentage": 5.92
                    }
                  ],
                  "total_keywords": 20,
                  "total_word_frequency": 3224,
                  "data_freshness": "current_cycle_only",
                  "note": "Keywords from current scraping cycle (refreshed every 4-5 hours)",
                  "timestamp": "2025-06-15T16:15:77.978597"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/jobs/search": {
      "get": {
        "summary": "Search and analyze jobs containing specific keyword from current scraping cycle",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "description": "Keyword to search for",
            "schema": {
              "type": "string",
              "minLength": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Keyword search analytics",
            "content": {
              "application/json": {
                "example": {
                  "keyword": "developer",
                  "total_matches": 174,
                  "unique_companies": 130,
                  "unique_sources": 12,
                  "match_percentage": 4.0,
                  "total_jobs_in_cycle": 4353,
                  "top_companies": [
                    {
                      "company": "Kapital Bank",
                      "job_count": 13
                    }
                  ],
                  "sources": [
                    {
                      "source": "Djinni",
                      "job_count": 126
                    }
                  ],
                  "data_freshness": "current_cycle_only",
                  "note": "Search results from current scraping cycle (refreshed every 4-5 hours)",
                  "timestamp": "2025-06-15T16:15:38.047965"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}